Mocking up Section 5
Please note there is no validation on these pages, and most of it is
copy and pasted to be as true to the paper form as possible.

Adding new pages to a11y.json

Removing snippets file

Changed all required to false

I'm not sure conditions are all required, but if you have a condition
the data probably is required.