Fixed unsafe regex
The regex for the telephone could apparantely cause a (catastrophic
backtracking)[https://www.regular-expressions.info/catastrophic.html]. I
couldn't figure out how to do it, but the linter thinks it could.

It looked like it was a result of the second `?` being greedy, so I just
appended another `?` so it wasn't greedy anymore.